# openapi.json
{"openapi":"3.1.0","info":{"title":"OpenAPI definition","version":"v0"},"servers":[{"url":"http://localhost:8080","description":"Generated server url"}],"paths":{"/api/v1/questoes":{"get":{"tags":["questao-controller"],"operationId":"listar","parameters":[{"name":"texto","in":"query","required":false,"schema":{"type":"string"}},{"name":"disciplina","in":"query","required":false,"schema":{"type":"string"}},{"name":"disciplinaId","in":"query","required":false,"schema":{"type":"integer","format":"int64"}},{"name":"assunto","in":"query","required":false,"schema":{"type":"string"}},{"name":"assuntoId","in":"query","required":false,"schema":{"type":"integer","format":"int64"}},{"name":"banca","in":"query","required":false,"schema":{"type":"string"}},{"name":"bancaId","in":"query","required":false,"schema":{"type":"integer","format":"int64"}},{"name":"instituicao","in":"query","required":false,"schema":{"type":"string"}},{"name":"instituicaoId","in":"query","required":false,"schema":{"type":"integer","format":"int64"}},{"name":"ano","in":"query","required":false,"schema":{"type":"integer","format":"int32"}},{"name":"cargo","in":"query","required":false,"schema":{"type":"string"}},{"name":"nivel","in":"query","required":false,"schema":{"type":"string"}},{"name":"modalidade","in":"query","required":false,"schema":{"type":"string"}},{"name":"sort","in":"query","required":false,"schema":{"type":"string"}},{"name":"page","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":0}},{"name":"size","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":10}}],"responses":{"401":{"description":"Unauthorized","content":{"application/problem+json":{"schema":{"$ref":"#/components/schemas/ProblemDetail"}}}},"403":{"description":"Forbidden","content":{"application/problem+json":{"schema":{"$ref":"#/components/schemas/ProblemDetail"}}}},"500":{"description":"Internal Server Error","content":{"application/problem+json":{"schema":{"$ref":"#/components/schemas/ProblemDetail"}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PagedModelQuestaoResponse"}}}}}},"post":{"tags":["questao-controller"],"operationId":"cadastrar","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/QuestaoRequest"}}},"required":true},"responses":{"401":{"description":"Unauthorized","content":{"application/problem+json":{"schema":{"$ref":"#/components/schemas/ProblemDetail"}}}},"403":{"description":"Forbidden","content":{"application/problem+json":{"schema":{"$ref":"#/components/schemas/ProblemDetail"}}}},"500":{"description":"Internal Server Error","content":{"application/problem+json":{"schema":{"$ref":"#/components/schemas/ProblemDetail"}}}},"201":{"description":"Created","content":{"*/*":{"schema":{"$ref":"#/components/schemas/QuestaoResponse"}}}}}}},"/api/v1/auth/register":{"post":{"tags":["auth-controller"],"operationId":"register","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RegisterRequest"}}},"required":true},"responses":{"401":{"description":"Unauthorized","content":{"application/problem+json":{"schema":{"$ref":"#/components/schemas/ProblemDetail"}}}},"403":{"description":"Forbidden","content":{"application/problem+json":{"schema":{"$ref":"#/components/schemas/ProblemDetail"}}}},"500":{"description":"Internal Server Error","content":{"application/problem+json":{"schema":{"$ref":"#/components/schemas/ProblemDetail"}}}},"201":{"description":"Created"}}}},"/api/v1/auth/login":{"post":{"tags":["auth-controller"],"operationId":"login","parameters":[{"name":"req","in":"query","required":true,"schema":{"$ref":"#/components/schemas/LoginRequest"}}],"responses":{"401":{"description":"Unauthorized","content":{"application/problem+json":{"schema":{"$ref":"#/components/schemas/ProblemDetail"}}}},"403":{"description":"Forbidden","content":{"application/problem+json":{"schema":{"$ref":"#/components/schemas/ProblemDetail"}}}},"500":{"description":"Internal Server Error","content":{"application/problem+json":{"schema":{"$ref":"#/components/schemas/ProblemDetail"}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/AuthResponse"}}}}}}},"/api/v1/admin/catalogo/subassuntos":{"post":{"tags":["sub-assunto-admin-controller"],"operationId":"cadastrar_1","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SubAssuntoRequest"}}},"required":true},"responses":{"401":{"description":"Unauthorized","content":{"application/problem+json":{"schema":{"$ref":"#/components/schemas/ProblemDetail"}}}},"403":{"description":"Forbidden","content":{"application/problem+json":{"schema":{"$ref":"#/components/schemas/ProblemDetail"}}}},"500":{"description":"Internal Server Error","content":{"application/problem+json":{"schema":{"$ref":"#/components/schemas/ProblemDetail"}}}},"201":{"description":"Created","content":{"*/*":{"schema":{"$ref":"#/components/schemas/CatalogoItemResponse"}}}}}}},"/api/v1/admin/catalogo/instituicoes":{"get":{"tags":["instituicao-admin-controller"],"operationId":"listar_1","responses":{"401":{"description":"Unauthorized","content":{"application/problem+json":{"schema":{"$ref":"#/components/schemas/ProblemDetail"}}}},"403":{"description":"Forbidden","content":{"application/problem+json":{"schema":{"$ref":"#/components/schemas/ProblemDetail"}}}},"500":{"description":"Internal Server Error","content":{"application/problem+json":{"schema":{"$ref":"#/components/schemas/ProblemDetail"}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/CatalogoItemResponse"}}}}}}},"post":{"tags":["instituicao-admin-controller"],"operationId":"cadastrar_2","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/InstituicaoRequest"}}},"required":true},"responses":{"401":{"description":"Unauthorized","content":{"application/problem+json":{"schema":{"$ref":"#/components/schemas/ProblemDetail"}}}},"403":{"description":"Forbidden","content":{"application/problem+json":{"schema":{"$ref":"#/components/schemas/ProblemDetail"}}}},"500":{"description":"Internal Server Error","content":{"application/problem+json":{"schema":{"$ref":"#/components/schemas/ProblemDetail"}}}},"201":{"description":"Created","content":{"*/*":{"schema":{"$ref":"#/components/schemas/CatalogoItemResponse"}}}}}}},"/api/v1/admin/catalogo/disciplinas":{"post":{"tags":["disciplina-admin-controller"],"operationId":"cadastrar_3","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/DisciplinaRequest"}}},"required":true},"responses":{"401":{"description":"Unauthorized","content":{"application/problem+json":{"schema":{"$ref":"#/components/schemas/ProblemDetail"}}}},"403":{"description":"Forbidden","content":{"application/problem+json":{"schema":{"$ref":"#/components/schemas/ProblemDetail"}}}},"500":{"description":"Internal Server Error","content":{"application/problem+json":{"schema":{"$ref":"#/components/schemas/ProblemDetail"}}}},"201":{"description":"Created","content":{"*/*":{"schema":{"$ref":"#/components/schemas/CatalogoItemResponse"}}}}}}},"/api/v1/admin/catalogo/bancas":{"post":{"tags":["admin-banca-controller"],"operationId":"cadastrar_4","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/BancaRequest"}}},"required":true},"responses":{"401":{"description":"Unauthorized","content":{"application/problem+json":{"schema":{"$ref":"#/components/schemas/ProblemDetail"}}}},"403":{"description":"Forbidden","content":{"application/problem+json":{"schema":{"$ref":"#/components/schemas/ProblemDetail"}}}},"500":{"description":"Internal Server Error","content":{"application/problem+json":{"schema":{"$ref":"#/components/schemas/ProblemDetail"}}}},"201":{"description":"Created","content":{"*/*":{"schema":{"$ref":"#/components/schemas/CatalogoItemResponse"}}}}}}},"/api/v1/admin/catalogo/assuntos":{"post":{"tags":["assunto-admin-controller"],"operationId":"cadastrar_5","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AssuntoRequest"}}},"required":true},"responses":{"401":{"description":"Unauthorized","content":{"application/problem+json":{"schema":{"$ref":"#/components/schemas/ProblemDetail"}}}},"403":{"description":"Forbidden","content":{"application/problem+json":{"schema":{"$ref":"#/components/schemas/ProblemDetail"}}}},"500":{"description":"Internal Server Error","content":{"application/problem+json":{"schema":{"$ref":"#/components/schemas/ProblemDetail"}}}},"201":{"description":"Created","content":{"*/*":{"schema":{"$ref":"#/components/schemas/CatalogoItemResponse"}}}}}}},"/api/v1/admin/usuarios/{id}/ativar":{"patch":{"tags":["admin-usuario-controller"],"operationId":"ativar","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"401":{"description":"Unauthorized","content":{"application/problem+json":{"schema":{"$ref":"#/components/schemas/ProblemDetail"}}}},"403":{"description":"Forbidden","content":{"application/problem+json":{"schema":{"$ref":"#/components/schemas/ProblemDetail"}}}},"500":{"description":"Internal Server Error","content":{"application/problem+json":{"schema":{"$ref":"#/components/schemas/ProblemDetail"}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/UsuarioPublicoResponse"}}}}}}},"/api/v1/questoes/{idQuestion}":{"get":{"tags":["questao-controller"],"operationId":"buscar","parameters":[{"name":"idQuestion","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"401":{"description":"Unauthorized","content":{"application/problem+json":{"schema":{"$ref":"#/components/schemas/ProblemDetail"}}}},"403":{"description":"Forbidden","content":{"application/problem+json":{"schema":{"$ref":"#/components/schemas/ProblemDetail"}}}},"500":{"description":"Internal Server Error","content":{"application/problem+json":{"schema":{"$ref":"#/components/schemas/ProblemDetail"}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/QuestaoResponse"}}}}}}},"/api/v1/catalogo/instituicoes":{"get":{"tags":["catalogo-instituicao-controller"],"operationId":"listar_2","responses":{"401":{"description":"Unauthorized","content":{"application/problem+json":{"schema":{"$ref":"#/components/schemas/ProblemDetail"}}}},"403":{"description":"Forbidden","content":{"application/problem+json":{"schema":{"$ref":"#/components/schemas/ProblemDetail"}}}},"500":{"description":"Internal Server Error","content":{"application/problem+json":{"schema":{"$ref":"#/components/schemas/ProblemDetail"}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Instituicao"}}}}}}}},"/api/v1/catalogo/disciplinas":{"get":{"tags":["disciplina-controller"],"operationId":"listar_3","responses":{"401":{"description":"Unauthorized","content":{"application/problem+json":{"schema":{"$ref":"#/components/schemas/ProblemDetail"}}}},"403":{"description":"Forbidden","content":{"application/problem+json":{"schema":{"$ref":"#/components/schemas/ProblemDetail"}}}},"500":{"description":"Internal Server Error","content":{"application/problem+json":{"schema":{"$ref":"#/components/schemas/ProblemDetail"}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/CatalogoItemResponse"}}}}}}}},"/api/v1/catalogo/disciplinas/{disciplinaId}/assuntos":{"get":{"tags":["assunto-controller"],"operationId":"listar_4","parameters":[{"name":"disciplinaId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"401":{"description":"Unauthorized","content":{"application/problem+json":{"schema":{"$ref":"#/components/schemas/ProblemDetail"}}}},"403":{"description":"Forbidden","content":{"application/problem+json":{"schema":{"$ref":"#/components/schemas/ProblemDetail"}}}},"500":{"description":"Internal Server Error","content":{"application/problem+json":{"schema":{"$ref":"#/components/schemas/ProblemDetail"}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/CatalogoItemResponse"}}}}}}}},"/api/v1/catalogo/bancas":{"get":{"tags":["catalogo-banca-controller"],"operationId":"listar_5","responses":{"401":{"description":"Unauthorized","content":{"application/problem+json":{"schema":{"$ref":"#/components/schemas/ProblemDetail"}}}},"403":{"description":"Forbidden","content":{"application/problem+json":{"schema":{"$ref":"#/components/schemas/ProblemDetail"}}}},"500":{"description":"Internal Server Error","content":{"application/problem+json":{"schema":{"$ref":"#/components/schemas/ProblemDetail"}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/CatalogoItemResponse"}}}}}}}},"/api/v1/catalogo/assuntos/{assuntoId}/subassuntos":{"get":{"tags":["sub-assunto-controller"],"operationId":"listar_6","parameters":[{"name":"assuntoId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"401":{"description":"Unauthorized","content":{"application/problem+json":{"schema":{"$ref":"#/components/schemas/ProblemDetail"}}}},"403":{"description":"Forbidden","content":{"application/problem+json":{"schema":{"$ref":"#/components/schemas/ProblemDetail"}}}},"500":{"description":"Internal Server Error","content":{"application/problem+json":{"schema":{"$ref":"#/components/schemas/ProblemDetail"}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/CatalogoItemResponse"}}}}}}}},"/api/v1/admin/questoes/{idQuestion}/gabarito":{"get":{"tags":["admin-questao-controller"],"operationId":"buscarGabarito","parameters":[{"name":"idQuestion","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"401":{"description":"Unauthorized","content":{"application/problem+json":{"schema":{"$ref":"#/components/schemas/ProblemDetail"}}}},"403":{"description":"Forbidden","content":{"application/problem+json":{"schema":{"$ref":"#/components/schemas/ProblemDetail"}}}},"500":{"description":"Internal Server Error","content":{"application/problem+json":{"schema":{"$ref":"#/components/schemas/ProblemDetail"}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/GabaritoResponse"}}}}}}}},"components":{"schemas":{"ProblemDetail":{"type":"object","properties":{"type":{"type":"string","format":"uri"},"title":{"type":"string"},"status":{"type":"integer","format":"int32"},"detail":{"type":"string"},"instance":{"type":"string","format":"uri"},"properties":{"type":"object","additionalProperties":{}}}},"QuestaoRequest":{"type":"object","properties":{"enunciado":{"type":"string","maxLength":10000,"minLength":0},"questao":{"type":"string","maxLength":10000,"minLength":0},"alternativas":{"type":"string","maxLength":10000,"minLength":0},"disciplina":{"type":"string","maxLength":160,"minLength":0},"assunto":{"type":"string","maxLength":200,"minLength":0},"banca":{"type":"string","maxLength":160,"minLength":0},"instituicao":{"type":"string","maxLength":200,"minLength":0},"disciplinaId":{"type":"integer","format":"int64"},"assuntoId":{"type":"integer","format":"int64"},"bancaId":{"type":"integer","format":"int64"},"instituicaoId":{"type":"integer","format":"int64"},"ano":{"type":"integer","format":"int32","maximum":2100,"minimum":1900},"cargo":{"type":"string","maxLength":160,"minLength":0},"nivel":{"type":"string","maxLength":80,"minLength":0},"modalidade":{"type":"string","minLength":1,"pattern":"^(A_E|A_D|CERTO_ERRADO)$"},"gabarito":{"type":"string","minLength":1}},"required":["alternativas","ano","assunto","banca","cargo","disciplina","enunciado","gabarito","instituicao","modalidade","nivel","questao"]},"QuestaoResponse":{"type":"object","properties":{"idQuestion":{"type":"string"},"enunciado":{"type":"string"},"questao":{"type":"string"},"alternativas":{"type":"string"},"disciplina":{"type":"string"},"disciplinaId":{"type":"integer","format":"int64"},"assunto":{"type":"string"},"assuntoId":{"type":"integer","format":"int64"},"banca":{"type":"string"},"bancaId":{"type":"integer","format":"int64"},"instituicao":{"type":"string"},"instituicaoId":{"type":"integer","format":"int64"},"ano":{"type":"integer","format":"int32"},"cargo":{"type":"string"},"nivel":{"type":"string"},"modalidade":{"type":"string"},"gabarito":{"type":"string"},"criadoEm":{"type":"string","format":"date-time"}}},"RegisterRequest":{"type":"object","properties":{"nome":{"type":"string","maxLength":160,"minLength":0},"email":{"type":"string","format":"email","maxLength":200,"minLength":0},"senha":{"type":"string","maxLength":200,"minLength":6}},"required":["email","nome","senha"]},"LoginRequest":{"type":"object","properties":{"email":{"type":"string","format":"email","maxLength":200,"minLength":0},"senha":{"type":"string","maxLength":200,"minLength":6}},"required":["email","senha"]},"AuthResponse":{"type":"object","properties":{"token":{"type":"string"},"email":{"type":"string"},"role":{"type":"string"}}},"SubAssuntoRequest":{"type":"object","properties":{"assuntoId":{"type":"integer","format":"int64"},"nome":{"type":"string","maxLength":200,"minLength":0}},"required":["assuntoId","nome"]},"CatalogoItemResponse":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"nome":{"type":"string"}}},"InstituicaoRequest":{"type":"object","properties":{"nome":{"type":"string","maxLength":200,"minLength":0}},"required":["nome"]},"DisciplinaRequest":{"type":"object","properties":{"nome":{"type":"string","maxLength":200,"minLength":0}},"required":["nome"]},"BancaRequest":{"type":"object","properties":{"nome":{"type":"string","maxLength":160,"minLength":0}},"required":["nome"]},"AssuntoRequest":{"type":"object","properties":{"disciplinaId":{"type":"integer","format":"int64"},"nome":{"type":"string","maxLength":200,"minLength":0}},"required":["disciplinaId","nome"]},"UsuarioPublicoResponse":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"nome":{"type":"string"},"email":{"type":"string"},"role":{"type":"string","enum":["ADMIN","VISITANTE"]},"status":{"type":"string","enum":["PENDENTE","ATIVO"]}}},"PageMetadata":{"type":"object","properties":{"size":{"type":"integer","format":"int64"},"number":{"type":"integer","format":"int64"},"totalElements":{"type":"integer","format":"int64"},"totalPages":{"type":"integer","format":"int64"}}},"PagedModelQuestaoResponse":{"type":"object","properties":{"content":{"type":"array","items":{"$ref":"#/components/schemas/QuestaoResponse"}},"page":{"$ref":"#/components/schemas/PageMetadata"}}},"Instituicao":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"nome":{"type":"string"}}},"GabaritoResponse":{"type":"object","properties":{"idQuestion":{"type":"string"},"gabarito":{"type":"string"}}}}}}